Edit the default args and import inspect
