+++ src/otx/core/model/module/action_classification.py
@@ -4,12 +4,11 @@
 """Class definition for action classification lightning module used in OTX."""
 from __future__ import annotations
 
-from functools import partial
+import inspect
 from typing import TYPE_CHECKING
 
 import torch
 from torch import Tensor
-from torchmetrics.classification.accuracy import Accuracy
 
 from otx.core.data.entity.action_classification import (
     ActionClsBatchDataEntity,
@@ -20,7 +19,9 @@ from otx.core.model.module.base import OTXLitModule
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
-    from torchmetrics import Metric
+    from torchmetrics.classification.accuracy import Accuracy
+
+    from otx.algo.metrices import MetricCallable
 
 
 class OTXActionClsLitModule(OTXLitModule):
@@ -32,7 +33,7 @@ class OTXActionClsLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = Accuracy,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -42,30 +43,15 @@ class OTXActionClsLitModule(OTXLitModule):
             metric=metric,
         )
 
-        self.metric = (
-            metric(
-                task="multiclass",
-                num_classes=self.model.num_classes,
-            )
-            if isinstance(metric, partial)
-            else metric
-        )
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                param_dict[name] = param.default if name != "num_classes" else self.model.num_classes
+            param_dict.pop("kwargs")
 
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+            metric = metric(**param_dict)  # type: ignore[call-arg]
+        self.metric = metric
 
     def _log_metrics(self, meter: Accuracy, key: str) -> None:
         results = meter.compute()
@@ -87,9 +73,10 @@ class OTXActionClsLitModule(OTXLitModule):
         if not isinstance(preds, ActionClsBatchPredEntity):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -115,6 +102,7 @@ class OTXActionClsLitModule(OTXLitModule):
         if not isinstance(preds, ActionClsBatchPredEntity):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
