+++ src/otx/core/model/module/base.py
@@ -37,7 +37,7 @@ class OTXLitModule(LightningModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric,
+        metric: Metric | None = None,
     ):
         super().__init__()
 
@@ -86,6 +86,26 @@ class OTXLitModule(LightningModule):
             return total_train_loss
 
         raise TypeError(train_loss)
+
+    def on_validation_epoch_start(self) -> None:
+        """Callback triggered when the validation epoch starts."""
+        if self.metric:
+            self.metric.reset()
+
+    def on_test_epoch_start(self) -> None:
+        """Callback triggered when the test epoch starts."""
+        if self.metric:
+            self.metric.reset()
+
+    def on_validation_epoch_end(self) -> None:
+        """Callback triggered when the validation epoch ends."""
+        if self.metric:
+            self._log_metrics(self.metric, "val")
+
+    def on_test_epoch_end(self) -> None:
+        """Callback triggered when the test epoch ends."""
+        if self.metric:
+            self._log_metrics(self.metric, "test")
 
     def setup(self, stage: str) -> None:
         """Lightning hook that is called at the beginning of fit (train + validate), validate, test, or predict.
@@ -187,11 +207,6 @@ class OTXLitModule(LightningModule):
                 ckpt_meta_info.label_names,
             )
         return super().load_state_dict(state_dict, *args, **kwargs)
-
-    @property
-    def lr_scheduler_monitor_key(self) -> str:
-        """Metric name that the learning rate scheduler monitor."""
-        return self.scheduler[-1].monitor
 
     @property
     def label_info(self) -> LabelInfo:
