+++ src/otx/core/model/module/classification.py
@@ -4,13 +4,11 @@
 """Class definition for classification lightning module used in OTX."""
 from __future__ import annotations
 
-from functools import partial
+import inspect
 from typing import TYPE_CHECKING
 
 import torch
 from torch import Tensor
-from torchmetrics.classification import MultilabelAccuracy
-from torchmetrics.classification.accuracy import Accuracy
 
 from otx.algo.classification.metrics import HLabelAccuracy
 from otx.core.data.dataset.classification import HLabelMetaInfo
@@ -30,8 +28,9 @@ from otx.core.model.module.base import OTXLitModule
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
-    from torchmetrics import Metric
+    from torchmetrics.classification.accuracy import Accuracy
 
+    from otx.algo.metrices import MetricCallable
     from otx.core.data.dataset.base import LabelInfo
 
 
@@ -44,7 +43,7 @@ class OTXMulticlassClsLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = Accuracy,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -54,30 +53,16 @@ class OTXMulticlassClsLitModule(OTXLitModule):
             metric=metric,
         )
 
-        self.metric = (
-            metric(
-                task="multiclass",
-                num_classes=self.model.num_classes,
-            )
-            if isinstance(metric, partial)
-            else metric
-        )
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                param_dict[name] = param.default if name != "num_classes" else self.model.num_classes
+            param_dict.pop("kwargs")
 
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
+            metric = metric(**param_dict)  # type: ignore[call-arg]
 
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+        self.metric = metric
 
     def _log_metrics(self, meter: Accuracy, key: str) -> None:
         results = meter.compute()
@@ -98,9 +83,11 @@ class OTXMulticlassClsLitModule(OTXLitModule):
 
         if not isinstance(preds, (MulticlassClsBatchPredEntity, MulticlassClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -126,9 +113,10 @@ class OTXMulticlassClsLitModule(OTXLitModule):
         if not isinstance(preds, (MulticlassClsBatchPredEntity, MulticlassClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
 
 class OTXMultilabelClsLitModule(OTXLitModule):
@@ -140,7 +128,7 @@ class OTXMultilabelClsLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = MultilabelAccuracy,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -149,32 +137,15 @@ class OTXMultilabelClsLitModule(OTXLitModule):
             scheduler=scheduler,
             metric=metric,
         )
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                param_dict[name] = param.default if name != "num_labels" else self.model.num_classes
+            param_dict.pop("kwargs")
 
-        self.metric = (
-            metric(
-                num_labels=self.model.num_classes,
-                threshold=metric.keywords["threshold"],
-                average=metric.keywords["average"],
-            )
-            if callable(metric)
-            else metric
-        )
-
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+            metric = metric(**param_dict)  # type: ignore[call-arg]
+        self.metric = metric
 
     def _log_metrics(self, meter: Accuracy, key: str) -> None:
         results = meter.compute()
@@ -192,9 +163,10 @@ class OTXMultilabelClsLitModule(OTXLitModule):
         if not isinstance(preds, (MultilabelClsBatchPredEntity, MultilabelClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -218,9 +190,10 @@ class OTXMultilabelClsLitModule(OTXLitModule):
         if not isinstance(preds, (MultilabelClsBatchPredEntity, MultilabelClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
 
 class OTXHlabelClsLitModule(OTXLitModule):
@@ -232,7 +205,7 @@ class OTXHlabelClsLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = HLabelAccuracy,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -242,20 +215,23 @@ class OTXHlabelClsLitModule(OTXLitModule):
             metric=metric,
         )
 
-        self.metric = (
-            metric(
-                num_multiclass_heads=self.model.num_multiclass_heads,
-                num_multilabel_classes=self.model.num_multilabel_classes,
-            )
-            if callable(metric)
-            else metric
-        )
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                if name in ["num_multiclass_heads", "num_multilabel_classes"]:
+                    param_dict[name] = self.model.get(name)
+                else:
+                    param_dict[name] = param.default
+            param_dict.pop("kwargs")
+
+            metric = metric(**param_dict)  # type: ignore[call-arg]
 
         # Temporary, TODO (sungmanc)
         # OTX can't support the auto-configuration for the H-label classification
         # Therefore, default metric can be None and it cause the error
         # This is the workaround
-        if self.metric is None:
+        else:
             self.metric = HLabelAccuracy(
                 num_multiclass_heads=self.model.num_multiclass_heads,
                 num_multilabel_classes=self.model.num_multilabel_classes,
@@ -277,25 +253,10 @@ class OTXHlabelClsLitModule(OTXLitModule):
         self.num_multilabel_classes = self.hlabel_info.num_multilabel_classes
         self.num_singlelabel_classes = self.num_labels - self.num_multilabel_classes
 
-        self.metric.num_multiclass_heads = self.num_multiclass_heads
-        self.metric.num_multilabel_classes = self.num_multilabel_classes
-        self.metric.set_hlabel_accuracy_from_head_logits_info(self.hlabel_info.head_idx_to_logits_range)
-
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+        if self.metric:
+            self.metric.num_multiclass_heads = self.num_multiclass_heads
+            self.metric.num_multilabel_classes = self.num_multilabel_classes
+            self.metric.set_hlabel_accuracy_from_head_logits_info(self.hlabel_info.head_idx_to_logits_range)
 
     def _log_metrics(self, meter: Accuracy, key: str) -> None:
         results = meter.compute()
@@ -313,9 +274,10 @@ class OTXHlabelClsLitModule(OTXLitModule):
         if not isinstance(preds, (HlabelClsBatchPredEntity, HlabelClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -345,9 +307,10 @@ class OTXHlabelClsLitModule(OTXLitModule):
         if not isinstance(preds, (HlabelClsBatchPredEntity, HlabelClsBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     @property
     def meta_info(self) -> LabelInfo:
