+++ src/otx/core/model/module/detection.py
@@ -4,13 +4,12 @@
 """Class definition for detection lightning module used in OTX."""
 from __future__ import annotations
 
+import inspect
 import logging as log
-from functools import partial
 from typing import TYPE_CHECKING
 
 import torch
 from torch import Tensor
-from torchmetrics.detection.mean_ap import MeanAveragePrecision
 
 from otx.core.data.entity.detection import (
     DetBatchDataEntity,
@@ -22,7 +21,9 @@ from otx.core.model.module.base import OTXLitModule
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
-    from torchmetrics import Metric
+    from torchmetrics.detection.mean_ap import MeanAveragePrecision
+
+    from otx.algo.metrices import MetricCallable
 
 
 class OTXDetectionLitModule(OTXLitModule):
@@ -34,7 +35,7 @@ class OTXDetectionLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = MeanAveragePrecision,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -44,30 +45,15 @@ class OTXDetectionLitModule(OTXLitModule):
             metric=metric,
         )
 
-        self.metric = (
-            metric(
-                box_format=metric.keywords["box_format"],
-                iou_type=metric.keywords["iou_type"],
-            )
-            if isinstance(metric, partial)
-            else metric
-        )
+        param_dict = {}
+        if metric:
+            sig = inspect.signature(metric)
+            for name, param in sig.parameters.items():
+                param_dict[name] = param.default
+            param_dict.pop("kwargs")
+            metric = metric(**param_dict)  # type: ignore[call-arg]
 
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+        self.metric = metric
 
     def _log_metrics(self, meter: MeanAveragePrecision, key: str) -> None:
         results = meter.compute()
@@ -102,9 +88,10 @@ class OTXDetectionLitModule(OTXLitModule):
         if not isinstance(preds, (DetBatchPredEntity, DetBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -145,6 +132,7 @@ class OTXDetectionLitModule(OTXLitModule):
         if not isinstance(preds, (DetBatchPredEntity, DetBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
