+++ src/otx/core/model/module/instance_segmentation.py
@@ -4,8 +4,8 @@
 """Class definition for instance segmentation lightning module used in OTX."""
 from __future__ import annotations
 
+import inspect
 import logging as log
-from functools import partial
 from typing import TYPE_CHECKING
 
 import torch
@@ -25,7 +25,8 @@ from otx.core.utils.mask_util import encode_rle, polygon_to_rle
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
-    from torchmetrics import Metric
+
+    from otx.algo.metrices import MetricCallable
 
 
 class OTXInstanceSegLitModule(OTXLitModule):
@@ -37,7 +38,7 @@ class OTXInstanceSegLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = OTXMaskRLEMeanAveragePrecision,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -46,32 +47,28 @@ class OTXInstanceSegLitModule(OTXLitModule):
             scheduler=scheduler,
             metric=metric,
         )
-        self.metric = (
-            metric(
-                box_format=metric.keywords["box_format"],
-                iou_type=metric.keywords["iou_type"],
-            )
-            if isinstance(metric, partial)
-            else metric
-        )
 
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                param_dict[name] = param.default
+            param_dict.pop("kwargs")
 
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
+            metric = metric(**param_dict)  # type: ignore[call-arg]
+        self.metric = metric
 
     def on_validation_epoch_end(self) -> None:
         """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-        self.metric.reset()
+        if self.metric:
+            self._log_metrics(self.metric, "val")
+            self.metric.reset()
 
     def on_test_epoch_end(self) -> None:
         """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
-        self.metric.reset()
+        if self.metric:
+            self._log_metrics(self.metric, "test")
+            self.metric.reset()
 
     def _log_metrics(self, meter: OTXMaskRLEMeanAveragePrecision, subset_name: str) -> None:
         results = meter.compute()
@@ -112,9 +109,10 @@ class OTXInstanceSegLitModule(OTXLitModule):
         if not isinstance(preds, (InstanceSegBatchPredEntity, InstanceSegBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -186,6 +184,7 @@ class OTXInstanceSegLitModule(OTXLitModule):
         if not isinstance(preds, (InstanceSegBatchPredEntity, InstanceSegBatchPredEntityWithXAI)):
             raise TypeError(preds)
 
-        self.metric.update(
-            **self._convert_pred_entity_to_compute_metric(preds, inputs),
-        )
+        if self.metric:
+            self.metric.update(
+                **self._convert_pred_entity_to_compute_metric(preds, inputs),
+            )
