+++ src/otx/core/model/module/segmentation.py
@@ -4,13 +4,12 @@
 """Class definition for segmentation lightning module used in OTX."""
 from __future__ import annotations
 
+import inspect
 import logging as log
-from functools import partial
 from typing import TYPE_CHECKING
 
 import torch
 from torch import Tensor
-from torchmetrics import Dice
 
 from otx.core.data.entity.segmentation import (
     SegBatchDataEntity,
@@ -22,7 +21,9 @@ from otx.core.model.module.base import OTXLitModule
 
 if TYPE_CHECKING:
     from lightning.pytorch.cli import LRSchedulerCallable, OptimizerCallable
-    from torchmetrics import Metric
+    from torchmetrics import Dice
+
+    from otx.algo.metrices import MetricCallable
 
 
 class OTXSegmentationLitModule(OTXLitModule):
@@ -34,7 +35,7 @@ class OTXSegmentationLitModule(OTXLitModule):
         torch_compile: bool,
         optimizer: list[OptimizerCallable] | OptimizerCallable = lambda p: torch.optim.SGD(p, lr=0.01),
         scheduler: list[LRSchedulerCallable] | LRSchedulerCallable = torch.optim.lr_scheduler.ConstantLR,
-        metric: Metric = Dice,
+        metric: MetricCallable | None = None,
     ):
         super().__init__(
             otx_model=otx_model,
@@ -49,30 +50,19 @@ class OTXSegmentationLitModule(OTXLitModule):
             Please, specify number of classes in config."""
             raise RuntimeError(msg)
 
-        self.metric = (
-            metric(
-                num_classes=self.model.num_classes + 1,
-                ignore_index=self.model.num_classes,
-            )
-            if isinstance(metric, partial)
-            else metric
-        )
+        if metric:
+            sig = inspect.signature(metric)
+            param_dict = {}
+            for name, param in sig.parameters.items():
+                if name == "num_classes":
+                    param_dict[name] = self.model.num_classes + 1
+                    param_dict["ignore_index"] = self.model.num_classes
+                else:
+                    param_dict[name] = param.default
+            param_dict.pop("kwargs")
 
-    def on_validation_epoch_start(self) -> None:
-        """Callback triggered when the validation epoch starts."""
-        self.metric.reset()
-
-    def on_test_epoch_start(self) -> None:
-        """Callback triggered when the test epoch starts."""
-        self.metric.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        """Callback triggered when the validation epoch ends."""
-        self._log_metrics(self.metric, "val")
-
-    def on_test_epoch_end(self) -> None:
-        """Callback triggered when the test epoch ends."""
-        self._log_metrics(self.metric, "test")
+            metric = metric(**param_dict)  # type: ignore[call-arg]
+        self.metric = metric
 
     def _log_metrics(self, meter: Dice, key: str) -> None:
         results = meter.compute()
@@ -106,8 +96,9 @@ class OTXSegmentationLitModule(OTXLitModule):
             raise TypeError(preds)
 
         predictions = self._convert_pred_entity_to_compute_metric(preds, inputs)
-        for prediction in predictions:
-            self.metric.update(**prediction)
+        if self.metric:
+            for prediction in predictions:
+                self.metric.update(**prediction)
 
     def _convert_pred_entity_to_compute_metric(
         self,
@@ -133,5 +124,6 @@ class OTXSegmentationLitModule(OTXLitModule):
         if not isinstance(preds, (SegBatchPredEntity, SegBatchPredEntityWithXAI)):
             raise TypeError(preds)
         predictions = self._convert_pred_entity_to_compute_metric(preds, inputs)
-        for prediction in predictions:
-            self.metric.update(**prediction)
+        if self.metric:
+            for prediction in predictions:
+                self.metric.update(**prediction)
